+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^14.4.3",
+    "axios": "^1.4.0",
     "react": "^18.2.0",
     "react-calendar": "^4.4.0",
     "react-dom": "^18.2.0",
