+++ src/index.js
@@ -2,10 +2,10 @@ import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { store } from './app/store';
+import { BrowserRouter } from 'react-router-dom';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import './index.css';
-import LoginPage from './components/loginSpace/login';
 
 const container = document.getElementById('root');
 const root = createRoot(container);
@@ -13,8 +13,10 @@ const root = createRoot(container);
 root.render(
   <React.StrictMode>
     <Provider store={store}>
-      {/* <App /> */}
-      <LoginPage />
+    <BrowserRouter>
+      <App />
+      {/* <LoginPage /> */}
+      </BrowserRouter>
     </Provider>
   </React.StrictMode>
 );
