+++ src/pages/mainPage/main.css
@@ -1,0 +1,3 @@
+.mainPage {
+    text-align: center;
+}
